arch(x86_64): complete 64-bit kernel porting, long mode paging, and interrupt pipeline
Migrate the Galio kernel architecture from 32-bit (i386) to 64-bit (x86_64).
Resolves boot stalls, early page faults, vector instruction faults, and 32-bit register truncations across kernel subsystems.

Summary of Changes:
- Boot & Paging:
  * Updated `boot.S` to set up long mode (PML4 -> PDPT -> PD) identity mapping for 4 GiB of physical memory using 2 MB pages.
  * Correctly initialized high 32 bits of 64-bit page directory entries.
  * Updated `linker.ld` to target `elf64-x86-64`.

- GDT & TSS Initialization:
  * Implemented 64-bit GDT descriptor reloading via far return (`retfq`) in `asm.s`.
  * Configured 64-bit Task State Segment (TSS) with `rsp0` kernel stack pointers for privilege transitions.

- Interrupt Handling & CPU State:
  * Refactored `registers_t` structure (`cpu.h`) to store 64-bit registers (`rax`-`r15`, `rip`, `rsp`, `rflags`).
  * Ported assembly ISR stubs (`isr_asm.s`) to save and restore 64-bit context frame.
  * Updated exception formatting in `isr.c` for 64-bit registers.

- Memory Management:
  * Removed legacy 32-bit CR0 manipulation from `paging.c`.
  * Fixed pointer truncation warnings (`u32` -> `uintptr_t`) in `heap.c`, `elf.c`, and `pmem.c`.
  * Updated `dma.c` and `e1000.c` to use 64-bit identity-mapped MMIO addresses.

- Process Management & Syscalls:
  * Replaced process register structures with 64-bit equivalents (`rip`, `rsp`, `rbp`, `rflags`).
  * Updated `process_switch_asm` to save/restore 64-bit callee-saved registers.
  * Corrected 64-bit `iretq` frame in `syscall_extra.c` for user privilege switching (User CS `0x1B`, User SS `0x23`).
  * Ported `syscall_handler` (`syscall.c`) to standard 64-bit calling conventions.

- Build System:
  * Added `-mno-sse -mno-sse2 -mno-mmx -mno-3dnow -mno-avx -mno-red-zone` to `CFLAGS` in `Makefile` to prevent kernel mode vector instruction crashes and stack corruption.

+++ kernel/mm/pmem.c
@@ -86,7 +86,7 @@ void pmem_init(u32 mmap_addr, u32 mmap_length) {
         while ((u32)entry < mmap_addr + mmap_length) {
             u32 addr = entry->addr_low;
             u32 len = entry->len_low;
-            u32 next = (u32)entry + entry->size + sizeof(entry->size);
+            u32 next = (uintptr_t)entry + entry->size + sizeof(entry->size);
 
             if (entry->type == MMAP_AVAILABLE && len > 0) {
                 u32 start_frame = addr / FRAME_SIZE;
